move demo stories to templates folder
